+++ pos_longpolling/models/pos_longpolling_models.py
@@ -8,18 +8,17 @@ class PosConfig(models.Model):
     _inherit = 'pos.config'
 
     # 5/60 = 0.0833 = 5 min - default value
-    query_timeout = fields.Float(string='Query timeout', default=0.0833,
-                                 help="Waiting period for any message from poll "
-                                      "(if we have not received a message at this period, "
-                                      "poll will send message ('PING') to check the connection)")
+    longpolling_max_silence_timeout = fields.Float(string='Max Silence timeout', default=0.0833,
+                                                   help="Waiting period for any message from poll "
+                                                        "(if we have not received a message at this period, "
+                                                        "poll will send message ('PING') to check the connection)")
 
     # 1/60 = 0.01666 = 1 min - default value
-    response_timeout = fields.Float(string='Response timeout', default=0.01666,
-                                    help="Waiting period for response message (i.e. once message from "
-                                         "poll has been sent, it will be waiting for response message ('PONG') "
-                                         "at this period and if the message has not been received, the icon turns "
-                                         "color to red. Once the connection is restored, the icon changes its color "
-                                         "back to green)")
+    longpolling_pong_timeout = fields.Float(string='Pong timeout', default=0.01666,
+                                            help="Waiting period to receive PONG message after sending PING request."
+                                                 "When this timeout occurs, the icon turns "
+                                                 "color to red. Once the connection is restored, the icon changes its color "
+                                                 "back to green)")
     autostart_longpolling = fields.Boolean('Autostart longpolling', default=True,
                                            help='When switched off longpoling will start only when other module start it')