+++ pos_mobile_restaurant/static/src/js/gui.js
@@ -10,10 +10,12 @@ odoo.define('pos_mobile_restaurant.gui', function (require) {
 
     gui.Gui.include({
         change_screen_type: function(current_screen) {
-            var swiper_floor_container = $('.swiper-floor-container');
+            var swiper_floor_container = $('.swiper-container-floor-screen');
             if (current_screen === "floors") {
                 swiper_floor_container.addClass('mobile-active-screen');
                 swiper_floor_container.css({display:''});
+                this.chrome.change_current_floor();
+
             } else {
                 swiper_floor_container.removeClass('mobile-active-screen');
                 swiper_floor_container.css({display:'none'});