+++ pos_product_category_discount/doc/changelog.rst
@@ -3,6 +3,16 @@
 Updates
 =======
 
+`1.2.2`
+------
+
+- **FIX:** The percentage discount information can be removed.
+
+`1.2.1`
+------
+
+- **NEW:** Support the POS Mobile UI for Waiters module
+
 `1.1.1`
 ------